requirements.txt
# Core Flask Framework
Flask==3.1.0
Werkzeug==3.1.3

# Database
mysql-connector-python==9.1.0
PyMySQL==1.1.1  # Alternative MySQL driver

# Security
Flask-Talisman==1.1.0  # Security headers
Flask-Limiter==3.8.0  # Rate limiting
Flask-WTF==1.2.2  # CSRF protection and forms
bcrypt==4.2.1  # Password hashing
cryptography==44.0.0  # Encryption utilities

# Excel/CSV Export
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0

# PDF Generation
reportlab==4.2.5
PyPDF2==3.0.1

# Image Processing
Pillow==11.0.0

# Email Support
Flask-Mail==0.10.0

# Environment Variables
python-dotenv==1.0.1

# Date/Time Utilities
python-dateutil==2.9.0

# Input Validation
email-validator==2.2.0
validators==0.34.0

# Task Queue (Optional for async operations)
celery==5.4.0
redis==5.2.1

# Logging
python-json-logger==3.2.1

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
